+++ Ejercicio 2/operations.c
@@ -1,0 +1,5 @@
+#include "operations.h"
+
+void calcular(tInfoCalc* pInfoCalc){
+    pInfoCalc->resultado = pInfoCalc->primero + pInfoCalc->segundo;
+}
