+++ Ejercicio 2/operations.h
@@ -1,0 +1,4 @@
+#include <stdio.h>
+#include "colaDinamica.h"
+
+void calcular(tInfoCalc* pInfoCalc);
