+++ Ejercicio 2/processOper.h
@@ -1,0 +1,7 @@
+#include <stdio.h>
+#include <string.h>
+#include "threadOper.h"
+
+extern int cargaPThread;
+
+int calcularRestantesEnProceso(FILE* filesrc, FILE* filedst, const int cargaPProcess);
